- done some edge cases

+++ backend/controllers/todoController.js
@@ -3,8 +3,17 @@ exports.createTodo = async (req ,res)=> {
     // verify the inputs
     try { 
         const {todo , category , date } = req.body; 
+        // check if the user is present
+        if(!req.user ) { 
+            throw new Error("Unauthorized To do this operation");
+        }
         if(!todo || !category || !date) { 
             throw new Error("Please Provide All Details")
+        }
+        // find the todo already exist ? 
+        const isTodoExited  = await Todo.findOne({ todo })
+        if(isTodoExited) { 
+            throw new Error("Todo Already Existed....")
         }
         // create todo 
         const todoDoc = Todo.create({ 
@@ -30,11 +39,17 @@ exports.deletedTodo = async (req ,res) => {
     try {
         // fetch the id from params 
         const {todoId} = req.params; 
+        if(!req.user ) { 
+            throw new Error("Unauthorized To do this operation");
+        }
         if(!todoId) { 
             throw new Error("Please Provide User Id")
         }
         // delete todo from database
         const deletedTodo = await Todo.findOneAndDelete({_id : todoId})
+        if(!deletedTodo) { 
+            throw new Error("Please Provide Valid Todo Id!")
+        }
         // return response
         return res.status(200).json( {
             success : true ,
@@ -55,24 +70,32 @@ exports.updateTodo = async (req ,res) => {
     try {
             
         const {todoId} = req.params; 
+        if(!req.user ) { 
+            throw new Error("Unauthorized To do this operation");
+        }
         const { todo , category , date} = req.body; 
-        if(!todo || !category || !date) { 
-            throw new Error("Please Enter All Details");
+        if(!todo && !date && !category) { 
+            throw new Error("Please Provide Some Values!")
         }
         if(!todoId) { 
             return res.status(404).json({ 
                 message : "Todo Id not found", 
             })
         }
-        // find todo 
+        // find todo is existed or not
+        const isTodoExited = await Todo.findById(todoId); 
+        if(!isTodoExited){ 
+            throw new Error("Please Provide Valid Todo Id!");
+        }
+
         const todoDoc = await Todo.findOneAndUpdate({ _id : todoId} , { 
             $set : { 
-                todo , 
-                category, 
-                date , 
+                todo : todo || isTodoExited.todo , 
+                category : category || isTodoExited.category, 
+                date : date || isTodoExited.date, 
             }, 
         },{new : true}) 
-        // 
+        
         return res.status(200).json( {
             success : true, 
             message : "Todo Updated Successfully" , 
@@ -86,31 +109,3 @@ exports.updateTodo = async (req ,res) => {
         })
     }
 }
-
-// exports.updateTodoTitle = async(req ,res)=> { 
-//     try { 
-//         const {todoId} = req.params; 
-//         const { todo}  = req.body; 
-//         if(!todo) { 
-//             throw new Error("Please Enter All Details"); 
-
-//         }
-//         if(!todoId)  { 
-//             return res.status(404).json({ 
-//                 success : false , 
-//                 message : "Todo Id not valid", 
-//             })
-//         }
-//         // 
-//         return res.status(200).json( {
-//             success : true, 
-//             message : "Todo Updated Successfully" , 
-//             todo : todoDoc,
-//         })
-//     }catch(err){ 
-//         return res.status(400).json({ 
-//             success : false, 
-//             message : err.message
-//         })
-//     }
-// }
